refactor: resolve comments
