+++ backend/grpc-server/src/utils.rs
@@ -119,3 +119,81 @@ fn parse_metadata<'a>(
             })
         })
 }
+
+#[macro_export]
+macro_rules! implement_payment_operation {
+    (
+        fn_name: $fn_name:ident,
+        log_prefix: $log_prefix:literal,
+        request_type: $request_type:ty,
+        response_type: $response_type:ty,
+        flow_marker: $flow_marker:ty,
+        resource_common_data_type: $resource_common_data_type:ty,
+        request_data_type: $request_data_type:ty,
+        response_data_type: $response_data_type:ty,
+        request_data_constructor: $request_data_constructor:path,
+        common_flow_data_constructor: $common_flow_data_constructor:path,
+        generate_response_fn: $generate_response_fn:path
+    ) => {
+        async fn $fn_name(
+            &self,
+            request: tonic::Request<$request_type>,
+        ) -> Result<tonic::Response<$response_type>, tonic::Status> {
+            tracing::info!(concat!($log_prefix, "_FLOW: initiated"));
+
+            let connector = $crate::utils::connector_from_metadata(request.metadata()).into_grpc_status()?;
+            let connector_auth_details = $crate::utils::auth_from_metadata(request.metadata()).into_grpc_status()?;
+            let payload = request.into_inner();
+
+            // Get connector data
+            let connector_data = connector_integration::types::ConnectorData::get_connector_by_name(&connector);
+
+            // Get connector integration
+            let connector_integration: hyperswitch_interfaces::connector_integration_v2::BoxedConnectorIntegrationV2<
+                '_,
+                $flow_marker,
+                $resource_common_data_type,
+                $request_data_type,
+                $response_data_type,
+            > = connector_data.connector.get_connector_integration_v2();
+
+            // Create connector request data
+            let specific_request_data = $request_data_constructor(payload.clone())
+                .into_grpc_status()?;
+
+            // Create common request data
+            let common_flow_data = $common_flow_data_constructor((payload.clone(), self.config.connectors.clone()))
+                .into_grpc_status()?;
+
+            // Create router data
+            let router_data = hyperswitch_domain_models::router_data_v2::RouterDataV2::<
+                $flow_marker,
+                $resource_common_data_type,
+                $request_data_type,
+                $response_data_type,
+            > {
+                flow: std::marker::PhantomData,
+                resource_common_data: common_flow_data,
+                connector_auth_type: connector_auth_details,
+                request: specific_request_data,
+                response: Err(hyperswitch_domain_models::router_data::ErrorResponse::default()),
+            };
+
+            // Execute connector processing
+            let response_result = external_services::service::execute_connector_processing_step(
+                &self.config.proxy,
+                connector_integration,
+                router_data,
+            )
+            .await
+            .switch()
+            .into_grpc_status()?;
+
+            // Generate response
+            let final_response = $generate_response_fn(response_result)
+                .into_grpc_status()?;
+
+            Ok(tonic::Response::new(final_response))
+        }
+    };
+}
